Created a separate UninstallCommand
